Renomeação das classes
