+++ pom.xml
@@ -19,5 +19,10 @@
   		<artifactId>mysql-connector-java</artifactId>
   		<version>5.1.26</version>
   	</dependency>
+  	<dependency>
+  		<groupId>junit</groupId>
+  		<artifactId>junit</artifactId>
+  		<version>4.12</version>
+  	</dependency>
   </dependencies>
 </project>
