+++ src/main/java/com/ssa/model/Area.java
@@ -1,4 +1,4 @@
-package com.ssa;
+package com.ssa.model;
 
 
 import java.io.Serializable;
