+++ src/main/java/com/ssa/model/DiarioBoletim.java
@@ -1,4 +1,4 @@
-package com.ssa;
+package com.ssa.model;
 
 import java.io.Serializable;
 import javax.persistence.Access;
@@ -14,7 +14,7 @@ import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 import javax.persistence.Table;
 
-import com.ssa.LiraRato.TipoImovel;
+import com.ssa.model.LiraRato.TipoImovel;
 
 import java.util.Date;
