+++ src/main/java/com/ssa/model/Imovel.java
@@ -1,4 +1,4 @@
-package com.ssa;
+package com.ssa.model;
 
 import java.io.Serializable;
 import java.util.List;
